+++ packages/ordia-core/ordia/__init__.py
@@ -18,11 +18,11 @@ try:
 
     except PackageNotFoundError:
 
-        __version__ = "0.8.0"
+        __version__ = "0.9.0"
 
 except ImportError:  # pragma: no cover
 
-    __version__ = "0.8.0"
+    __version__ = "0.9.0"